销售报价单

+++ src/main/java/com/accp/erp/service/impl/SmlordbillmainServiceImpl.java
@@ -5,7 +5,10 @@ import com.accp.erp.dao.SmlordbillsubDao;
 import com.accp.erp.entity.Smlordbillmain;
 import com.accp.erp.entity.Smlordbillsub;
 import com.accp.erp.service.ISmlordbillmainService;
+import com.baomidou.mybatisplus.core.conditions.Wrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 
 import java.text.ParseException;
@@ -47,7 +50,9 @@ public class SmlordbillmainServiceImpl extends ServiceImpl<SmlordbillmainDao, Sm
 	public boolean add(Smlordbillmain smlordbillmain) {
 		int bol=baseMapper.insert(smlordbillmain);
 		if(bol>0) {
-			for (Smlordbillsub s : smlordbillmain.getSubList()) {
+			for (int i = 0; i < smlordbillmain.getSubList().size(); i++) {
+				Smlordbillsub s=smlordbillmain.getSubList().get(i);
+				s.setRowNO(i+1);
 				s.setBillNo(smlordbillmain.getBillNo());
 				s.setFlag(smlordbillmain.getFlag());
 				s.setBillDate(smlordbillmain.getBillDate());
@@ -58,14 +63,54 @@ public class SmlordbillmainServiceImpl extends ServiceImpl<SmlordbillmainDao, Sm
 		return bol>0;
 	}
 	@Override
-	public boolean remove(String billNo) {
+	public String remove(String billNo,Integer flag) {
+		Smlordbillmain smlordbillmain=baseMapper.queryOne(billNo, flag);
+		if(smlordbillmain.getAuditStatus()==1) {
+			return "该订单已审核，不能被删除";
+		}
 		QueryWrapper<Smlordbillmain> queryWrapper=new QueryWrapper<>();
 		queryWrapper.eq(Smlordbillmain.BILLNO, billNo);
+		queryWrapper.eq(Smlordbillmain.FLAG, flag);
 		int bool=baseMapper.delete(queryWrapper);
 		QueryWrapper<Smlordbillsub> queryWrapper2=new QueryWrapper<>();
 		queryWrapper2.eq(Smlordbillsub.BILLNO, billNo);
+		queryWrapper2.eq(Smlordbillsub.FLAG, flag);
 		subDao.delete(queryWrapper2);
+		return "true";
+	}
+	
+	@Override
+	public boolean updateSmlordbillmain(Smlordbillmain smlordbillmain) {
+		QueryWrapper<Smlordbillmain> queryWrapper=new QueryWrapper<>();
+		queryWrapper.eq(Smlordbillmain.BILLNO, smlordbillmain.getBillNo());
+		queryWrapper.eq(Smlordbillmain.FLAG, smlordbillmain.getFlag());
+		int bool=baseMapper.update(smlordbillmain, queryWrapper);
+		if(bool>0) {
+			//删除明细表的数据
+			QueryWrapper<Smlordbillsub> queryWrapper2=new QueryWrapper<>();
+			queryWrapper2.eq(Smlordbillsub.BILLNO, smlordbillmain.getBillNo());
+			queryWrapper2.eq(Smlordbillsub.FLAG, smlordbillmain.getFlag());
+			subDao.delete(queryWrapper2);
+			//新增明细表的数据
+			for (int i = 0; i < smlordbillmain.getSubList().size(); i++) {
+				Smlordbillsub s=smlordbillmain.getSubList().get(i);
+				s.setRowNO(i+1);
+				s.setBillNo(smlordbillmain.getBillNo());
+				s.setFlag(smlordbillmain.getFlag());
+				s.setBillDate(smlordbillmain.getBillDate());
+				subDao.insert(s);
+			}
+		}
 		return bool>0;
+	}
+	@Override
+	public Smlordbillmain queryOne(String billNo, Integer flag) {
+		
+		return baseMapper.queryOne(billNo, flag);
+	}
+	@Override
+	public IPage<Smlordbillmain> select(Page page, Wrapper<Smlordbillmain> wrapper) {
+		return baseMapper.select(page, wrapper);
 	}
